change state to movies and users
