--- app/reducers/movie-reducer.js
@@ -1,1 +1,0 @@
-
